New scenario based on AEA MultiAgentManager
Added new scenario for marketplace, which is based on MultiAgentManager.
Contains a script to start different agents and overwrites
configurations like ports and current contract address to be used.
Readme is updated to describe how current scenario can be started.
Updated script in service-directory to dump contract address for new
scenario in a text file.
Updated gitignore to current state.

Signed-off-by: Daniel Kunz <daniel.kunz@de.bosch.com>